requirements-api.txt
-r requirements.txt

fastapi>=0.110,<0.120
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7,<2.0
bcrypt<4.1
python-multipart>=0.0.9
email-validator>=2.0
rq>=1.16
redis>=5.0
sse-starlette>=2.1
